Split hljs into a new file; improved dev vite server system
