+++ templates/base/base_footer.html
@@ -15,6 +15,7 @@
 </p>
 </div>
 
+<script type="module" src="{{ asset "hljs.ts" }}"></script>
 
 </div>
 </body>
